// src/charting_library.d.ts
export declare const widget: ChartingLibraryWidgetConstructor
// ...
export interface ChartingLibraryWidgetConstructor {
    new (options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions): IChartingLibraryWidget
}
export interface ChartingLibraryWidgetOptions {
    /** @deprecated */
    container_id?: string
    container: HTMLElement | string
    datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi)
    interval: ResolutionString
    symbol?: string
    auto_save_delay?: number
    autosize?: boolean
    debug?: boolean
    disabled_features?: string[]
    drawings_access?: AccessList
    enabled_features?: string[]
    fullscreen?: boolean
    height?: number
    library_path?: string
    locale: LanguageCode
    numeric_formatting?: NumericFormattingParams
    saved_data?: object
    saved_data_meta_info?: SavedStateMetaInfo
    studies_access?: AccessList
    study_count_limit?: number
    symbol_search_request_delay?: number
    timeframe?: string
    timezone?: "exchange" | Timezone
    toolbar_bg?: string
    width?: number
    charts_storage_url?: string
    charts_storage_api_version?: AvailableSaveloadVersions
    client_id?: string
    user_id?: string
    load_last_chart?: boolean
    studies_overrides?: StudyOverrides
    customFormatters?: CustomFormatters
    custom_formatters?: CustomFormatters
    overrides?: Overrides
    snapshot_url?: string
    preset?: "mobile"
    time_frames?: TimeFrameItem[]
    custom_css_url?: string
    favorites?: Favorites
    save_load_adapter?: IExternalSaveLoadAdapter
    loading_screen?: LoadingScreenOptions
    settings_adapter?: ISettingsAdapter
    theme?: ThemeName
    compare_symbols?: CompareSymbol[]
    custom_indicators_getter?: (PineJS: PineJS) => Promise<readonly CustomIndicator[]>
    additional_symbol_info_fields?: AdditionalSymbolInfoField[]
    header_widget_buttons_mode?: HeaderWidgetButtonsMode
    time_scale?: TimeScaleOptions
    symbol_search_complete?: SymbolSearchCompleteOverrideFunction
}
// ...
export interface IChartingLibraryWidget {
    headerReady(): Promise<void>
    onChartReady(callback: EmptyCallback): void
    onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void
    onShortcut(shortCut: string | number | (string | number)[], callback: EmptyCallback): void
    subscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void
    unsubscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void
    chart(index?: number): IChartWidgetApi
    getLanguage(): LanguageCode
    setSymbol(symbol: string, interval: ResolutionString, callback: EmptyCallback): void
    remove(): void
    closePopupsAndDialogs(): void
    selectLineTool(linetool: SupportedLineTools): void
    selectedLineTool(): SupportedLineTools
    save(callback: (state: object) => void): void
    load(state: object, extendedData?: SavedStateMetaInfo): void
    getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void
    loadChartFromServer(chartRecord: SaveLoadChartRecord): void
    saveChartToServer(onComplete?: EmptyCallback, onFail?: EmptyCallback, options?: SaveChartToServerOptions): void
    removeChartFromServer(chartId: string, onCompleteCallback: EmptyCallback): void
    onContextMenu(callback: (unixTime: number, price: number) => ContextMenuItem[]): void
    createButton(options?: CreateButtonOptions): HTMLElement
    createDropdown(params: DropdownParams): Promise<IDropdownApi>
    showNoticeDialog(params: DialogParams<() => void>): void
    showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void
    showLoadChartDialog(): void
    showSaveAsChartDialog(): void
    symbolInterval(): SymbolIntervalResult
    mainSeriesPriceFormatter(): INumberFormatter
    getIntervals(): string[]
    getStudiesList(): string[]
    getStudyInputs(studyName: string): StudyInputInfo[]
    addCustomCSSFile(url: string): void
    applyOverrides<TOverrides extends StudyOverrides>(overrides: TOverrides): void
    applyStudiesOverrides(overrides: object): void
    watchList(): Promise<IWatchListApi>
    news(): Promise<INewsApi>
    activeChart(): IChartWidgetApi
    chartsCount(): number
    layout(): LayoutType
    setLayout(layout: LayoutType): void
    layoutName(): string
    changeTheme(themeName: ThemeName, options?: ChangeThemeOptions): Promise<void>
    getTheme(): ThemeName
    takeScreenshot(): void
    takeClientScreenshot(options?: Partial<ClientSnapshotOptions>): Promise<HTMLCanvasElement>
    lockAllDrawingTools(): IWatchedValue<boolean>
    hideAllDrawingTools(): IWatchedValue<boolean>
    magnetEnabled(): IWatchedValue<boolean>
    magnetMode(): IWatchedValue<number>
    symbolSync(): IWatchedValue<boolean>
    intervalSync(): IWatchedValue<boolean>
    crosshairSync(): IWatchedValue<boolean>
    timeSync(): IWatchedValue<boolean>
    startFullscreen(): void
    exitFullscreen(): void
    undoRedoState(): UndoRedoState
    navigationButtonsVisibility(): IWatchedValue<VisibilityType>
    paneButtonsVisibility(): IWatchedValue<VisibilityType>
    dateFormat(): IWatchedValue<DateFormat>
}
// ...
export interface TradingTerminalWidgetOptions extends ChartingLibraryWidgetOptions {
    brokerConfig?: SingleBrokerMetaInfo
    broker_config?: SingleBrokerMetaInfo
    restConfig?: RestBrokerConnectionInfo
    widgetbar?: WidgetBarParams
    rss_news_feed?: RssNewsFeedParams
    news_provider?: GetNewsFunction
    trading_customization?: TradingCustomization
    brokerFactory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal
    broker_factory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal
}